requirements.txt
streamlit
google-generativeai
pyttsx3